MachineLearningModule/django/requirements.txt
asgiref==3.3.4
boto3==1.17.62
botocore==1.20.62
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
jmespath==0.10.0
joblib==1.0.1
numpy==1.20.2
pandas==1.2.4
Pygments==2.8.1
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
urllib3==1.26.4
